requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.6.0
    # via
    #   -r .\requirements.in
    #   django
build==0.9.0
    # via
    #   -r .\requirements.in
    #   pip-tools
cffi==1.15.1
    # via
    #   -r .\requirements.in
    #   cryptography
click==8.1.3
    # via
    #   -r .\requirements.in
    #   pip-tools
colorama==0.4.6
    # via
    #   -r .\requirements.in
    #   build
    #   click
cryptography==39.0.0
    # via
    #   -r .\requirements.in
    #   django-rest-knox
django==4.1.5
    # via
    #   -r .\requirements.in
    #   django-rest-knox
    #   djangorestframework
django-rest-knox==4.2.0
    # via -r .\requirements.in
djangorestframework==3.14.0
    # via
    #   -r .\requirements.in
    #   django-rest-knox
flake8==6.0.0
    # via -r .\requirements.in
mccabe==0.7.0
    # via
    #   -r .\requirements.in
    #   flake8
packaging==23.0
    # via
    #   -r .\requirements.in
    #   build
pep517==0.13.0
    # via
    #   -r .\requirements.in
    #   build
pip-tools==6.12.1
    # via -r .\requirements.in
psycopg2==2.9.5
    # via -r .\requirements.in
pycodestyle==2.10.0
    # via
    #   -r .\requirements.in
    #   flake8
pycparser==2.21
    # via
    #   -r .\requirements.in
    #   cffi
pyflakes==3.0.1
    # via
    #   -r .\requirements.in
    #   flake8
python-decouple==3.7
    # via -r .\requirements.in
pytz==2022.7
    # via
    #   -r .\requirements.in
    #   djangorestframework
sqlparse==0.4.3
    # via
    #   -r .\requirements.in
    #   django
tzdata==2022.7
    # via
    #   -r .\requirements.in
    #   django
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
